removes some compilation warnings.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@1846 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/jackal/jccl/Config/VarValue.cpp
@@ -411,7 +411,7 @@ ostream& operator << (ostream& out, const vjVarValue& v) {
 	    out << *(v.embeddedchunkval);
 	return out;
     default:
-	out << "[can't print value for type " << v.type << " ]";
+	out << "[can't print value for type " << (int)v.type << " ]";
 	return out;
     }
 }
